+++ src/Components/TodoCounter/index.js
@@ -1,10 +1,15 @@
 import React from 'react'
+import { TodoContext } from '../TodoContext'
 import './TodoCounter.css'
 
-function TodoCounter({total, completed}) {
+function TodoCounter() {
+  const {
+    completedTodos,
+    totalTodos,
+  } = React.useContext(TodoContext)
   return (
     <h1 className='TodoCounter'>
-      You have completed <span>{completed}</span> out of <span>{total}</span> TODOs
+      You have completed <span>{completedTodos}</span> out of <span>{totalTodos}</span> TODOs
     </h1>
   )
 }
