+++ src/Components/TodoSearch/index.js
@@ -1,20 +1,22 @@
 import React from 'react'
+import { TodoContext } from '../TodoContext'
 import './TodoSearch.css'
 
-function TodoSearch({
-  searchValue,
-  setSearchValue,
-}) 
+function TodoSearch(){
+  const {
+    searchValue,
+    setSearchValue,
+} = React.useContext(TodoContext) 
 
-{
-  return (
+return (
     <input
       className='TodoSearch'
       placeholder='Study' 
       value={searchValue}
       onChange={(event) => {
         setSearchValue(event.target.value)
-      }}/>
+      }}
+    />
   )
 }
